+++ src/app/queue.hpp
@@ -19,8 +19,8 @@ void QueueInit(Queue_t *q);
 u8 IsFull(Queue_t *q);
 u8 IsEmpty(Queue_t *q);
 void EnQueue(Queue_t *q, u8 *val, u8 lenth);
-int Pop(Queue_t *buffer1, Queue_t *buffer2, u8 data[11]);
+i8 Pop(Queue_t *buffer1, Queue_t *buffer2, u8 data[11]);
 void UI_EnQueue(Queue_t *q, u8 *val, u8 lenth);
-int UI_Pop(Queue_t *buffer1, u8 *data);
+i8 UI_Pop(Queue_t *buffer1, u8 *data);
 
 #endif /* __QUEUE_HPP__ */
